Crash Fix Related to PDF Viewer

+++ Chummer/frmViewer.cs
@@ -34,7 +34,6 @@ namespace Chummer
 		private XmlDocument _objCharacterXML = new XmlDocument();
 		private string _strSelectedSheet = "Shadowrun 5";
 		private bool _blnLoading = false;
-		
 		#region Control Events
 		public frmViewer()
 		{
@@ -53,58 +52,67 @@ namespace Chummer
 			List<ListItem> lstFiles = new List<ListItem>();
 			// Populate the XSLT list with all of the XSL files found in the sheets directory.
 			string sheetsDirectoryPath = Path.Combine(Application.StartupPath, "sheets");
-			foreach (string strFile in Directory.GetFiles(sheetsDirectoryPath))
-			{
-				// Only show files that end in .xsl. Do not include files that end in .xslt since they are used as "hidden" reference sheets (hidden because they are partial templates that cannot be used on their own).
-				if (!strFile.EndsWith(".xslt") && strFile.EndsWith(".xsl"))
-				{
-					string strFileName = Path.GetFileNameWithoutExtension(strFile);
-					ListItem objItem = new ListItem();
-					objItem.Value = strFileName;
-					objItem.Name = strFileName;
-					lstFiles.Add(objItem);
+		    if (Directory.Exists(sheetsDirectoryPath))
+		    {
+		        foreach (string strFile in Directory.GetFiles(sheetsDirectoryPath))
+		        {
+		            // Only show files that end in .xsl. Do not include files that end in .xslt since they are used as "hidden" reference sheets (hidden because they are partial templates that cannot be used on their own).
+		            if (!strFile.EndsWith(".xslt") && strFile.EndsWith(".xsl"))
+		            {
+		                string strFileName = Path.GetFileNameWithoutExtension(strFile);
+		                ListItem objItem = new ListItem();
+		                objItem.Value = strFileName;
+		                objItem.Name = strFileName;
+		                lstFiles.Add(objItem);
 
-					//cboXSLT.Items.Add(strFileName);
-				}
-			}
+		                //cboXSLT.Items.Add(strFileName);
+		            }
+		        }
+		    }
 
-            // Populate the XSL list with all of the XSL files found in the sheets\[language] directory.
+		    // Populate the XSL list with all of the XSL files found in the sheets\[language] directory.
             if (GlobalOptions.Instance.Language != "en-us")
             {
                 XmlDocument objLanguageDocument = LanguageManager.Instance.XmlDoc;
-                string strLanguage = objLanguageDocument.SelectSingleNode("/chummer/name").InnerText;
+                string strLanguage = objLanguageDocument.SelectSingleNode("/chummer/name")?.InnerText;
                 string languageDirectoryPath = Path.Combine(Application.StartupPath, "sheets", GlobalOptions.Instance.Language);
 
-                foreach (string strFile in Directory.GetFiles(languageDirectoryPath))
+                if (Directory.Exists(languageDirectoryPath))
                 {
-                    // Only show files that end in .xsl. Do not include files that end in .xslt since they are used as "hidden" reference sheets (hidden because they are partial templates that cannot be used on their own).
-                    if (!strFile.EndsWith(".xslt") && strFile.EndsWith(".xsl"))
+                    foreach (string strFile in Directory.GetFiles(languageDirectoryPath))
                     {
-                        string strFileName = Path.GetFileNameWithoutExtension(strFile);
-                        ListItem objItem = new ListItem();
-                        objItem.Value = Path.Combine(GlobalOptions.Instance.Language, strFileName);
-                        objItem.Name = strLanguage + ": " + strFileName;
-                        lstFiles.Add(objItem);
+                        // Only show files that end in .xsl. Do not include files that end in .xslt since they are used as "hidden" reference sheets (hidden because they are partial templates that cannot be used on their own).
+                        if (!strFile.EndsWith(".xslt") && strFile.EndsWith(".xsl"))
+                        {
+                            string strFileName = Path.GetFileNameWithoutExtension(strFile);
+                            ListItem objItem = new ListItem();
+                            objItem.Value = Path.Combine(GlobalOptions.Instance.Language, strFileName);
+                            objItem.Name = strLanguage + ": " + strFileName;
+                            lstFiles.Add(objItem);
+                        }
                     }
                 }
             }
 
             // Populate the XSLT list with all of the XSL files found in the sheets\omae directory.
             string omaeDirectoryPath = Path.Combine(Application.StartupPath, "sheets", "omae");
-            foreach (string strFile in Directory.GetFiles(omaeDirectoryPath))
-            {
-                // Only show files that end in .xsl. Do not include files that end in .xslt since they are used as "hidden" reference sheets (hidden because they are partial templates that cannot be used on their own).
-                if (!strFile.EndsWith(".xslt") && strFile.EndsWith(".xsl"))
-                {
-                    string strFileName = Path.GetFileNameWithoutExtension(strFile);
-                    ListItem objItem = new ListItem();
-                    objItem.Value = Path.Combine("omae", strFileName);
-                    objItem.Name = LanguageManager.Instance.GetString("Menu_Main_Omae") + ": " + strFileName;
-                    lstFiles.Add(objItem);
-                }
-            }
+		    if (Directory.Exists(omaeDirectoryPath))
+		    {
+		        foreach (string strFile in Directory.GetFiles(omaeDirectoryPath))
+		        {
+		            // Only show files that end in .xsl. Do not include files that end in .xslt since they are used as "hidden" reference sheets (hidden because they are partial templates that cannot be used on their own).
+		            if (!strFile.EndsWith(".xslt") && strFile.EndsWith(".xsl"))
+		            {
+		                string strFileName = Path.GetFileNameWithoutExtension(strFile);
+		                ListItem objItem = new ListItem();
+		                objItem.Value = Path.Combine("omae", strFileName);
+		                objItem.Name = LanguageManager.Instance.GetString("Menu_Main_Omae") + ": " + strFileName;
+		                lstFiles.Add(objItem);
+		            }
+		        }
+		    }
 
-            cboXSLT.ValueMember = "Value";
+		    cboXSLT.ValueMember = "Value";
 			cboXSLT.DisplayMember = "Name";
 			cboXSLT.DataSource = lstFiles;
 
@@ -139,11 +147,10 @@ namespace Chummer
 		private void cmdSaveHTML_Click(object sender, EventArgs e)
 		{
 			// Save the generated output as HTML.
-			string strSaveFile = string.Empty;
 			SaveFileDialog1.Filter = "HTML Page|*.htm";
 			SaveFileDialog1.Title = LanguageManager.Instance.GetString("Button_Viewer_SaveAsHtml");
 			SaveFileDialog1.ShowDialog();
-			strSaveFile = SaveFileDialog1.FileName;
+            string strSaveFile = SaveFileDialog1.FileName;
 
 			if (string.IsNullOrEmpty(strSaveFile))
 				return;
@@ -151,17 +158,15 @@ namespace Chummer
 			TextWriter objWriter = new StreamWriter(strSaveFile, false, Encoding.UTF8);
 			objWriter.Write(webBrowser1.DocumentText);
 			objWriter.Close();
-
 		}
 
 		private void cmdSaveXML_Click(object sender, EventArgs e)
 		{
 			// Save the printout XML generated by the character.
-			string strSaveFile = string.Empty;
 			SaveFileDialog1.Filter = "XML File|*.xml";
 			SaveFileDialog1.Title = LanguageManager.Instance.GetString("Button_Viewer_SaveAsXml");
 			SaveFileDialog1.ShowDialog();
-			strSaveFile = SaveFileDialog1.FileName;
+			string strSaveFile = SaveFileDialog1.FileName;
 
 			if (string.IsNullOrEmpty(strSaveFile))
 				return;
@@ -173,12 +178,15 @@ namespace Chummer
 		{
             // Remove the mugshots directory when the form closes.
             string mugshotsDirectoryPath = Path.Combine(Application.StartupPath, "mugshots");
-            try
+            if (Directory.Exists(mugshotsDirectoryPath))
 			{
-                Directory.Delete(mugshotsDirectoryPath, true);
-			}
-			catch
-			{
+			    try
+			    {
+                    Directory.Delete(mugshotsDirectoryPath, true);
+                }
+			    catch (IOException)
+			    {
+			    }
 			}
 
 			// Clear the reference to the character's Print window.
@@ -196,22 +204,30 @@ namespace Chummer
 		/// </summary>
 		private void GenerateOutput()
 		{
-			XslCompiledTransform objXSLTransform = new XslCompiledTransform();
-			try
-			{
-				objXSLTransform.Load(Path.Combine(Application.StartupPath, "sheets", _strSelectedSheet + ".xsl"));
-			}
-			catch(Exception ex)
-			{
-				string strReturn = string.Format("Error attempting to load {0}\n", _strSelectedSheet);
-				Log.Enter(strReturn);
-				Log.Error("ERROR Message = " + ex.Message);
-				strReturn += ex.Message;
-				MessageBox.Show(strReturn);
-				return;
-			}
+		    string strXslPath = Path.Combine(Application.StartupPath, "sheets", _strSelectedSheet + ".xsl");
+		    if (!File.Exists(strXslPath))
+		    {
+                string strReturn = string.Format("File not found when attempting to load {0}\n", _strSelectedSheet);
+                Log.Enter(strReturn);
+                MessageBox.Show(strReturn);
+                return;
+            }
+		    XslCompiledTransform objXSLTransform = new XslCompiledTransform();
+		    try
+		    {
+		        objXSLTransform.Load(strXslPath);
+		    }
+		    catch (Exception ex)
+		    {
+		        string strReturn = string.Format("Error attempting to load {0}\n", _strSelectedSheet);
+		        Log.Enter(strReturn);
+		        Log.Error("ERROR Message = " + ex.Message);
+		        strReturn += ex.Message;
+		        MessageBox.Show(strReturn);
+		        return;
+		    }
 
-			MemoryStream objStream = new MemoryStream();
+		    MemoryStream objStream = new MemoryStream();
 			XmlTextWriter objWriter = new XmlTextWriter(objStream, Encoding.UTF8);
 
 			objXSLTransform.Transform(_objCharacterXML, null, objWriter);
@@ -302,15 +318,14 @@ namespace Chummer
         private void cmdSavePDF_Click(object sender, EventArgs e)
         {
             // Save the generated output as PDF.
-            string strSaveFile = "";
             SaveFileDialog1.Filter = "PDF|*.pdf";
             SaveFileDialog1.Title = LanguageManager.Instance.GetString("Button_Viewer_SaveAsPdf");
             SaveFileDialog1.ShowDialog();
-            strSaveFile = SaveFileDialog1.FileName;
+            string strSaveFile = SaveFileDialog1.FileName;
 
             if (strSaveFile == "")
                 return;
-            var converter = new HtmlToPdf();
+            HtmlToPdf converter = new HtmlToPdf();
             converter.Options.MarginBottom = 10;
             converter.Options.MarginTop = 10;
             converter.Options.MarginLeft = 10;
